Preserve working directory across restarts (#6416)
This change allows the working directory to persist across restarts, so
that after a restart you don't need to re-navigate.

<img width="618" alt="image"
src="https://github.com/user-attachments/assets/ef859149-6ce3-4e38-951f-f34892d17c7d"
/>

Most of the work is in the supervisor; this is just the client bits. I
would have preferred for this change to _only_ exist in the supervisor,
but it turned out that on Windows there is no way to examine another
process's current working directory if it has been changed, so the
supervisor actually has no visibility into the working directory without
trying to sniff the UI comm traffic or something equally dreadful.

Since we already plumb the working directory into the front end in order
to display it in the Console, the approach taken here is to have the
front end tell the supervisor the directory in which to perform the
restart, since it already knows it.

Addresses #3188.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- The current working directory is now preserved across R and Python
restarts (#3188)

#### Bug Fixes

- N/A


### QA Notes

This should be tested on Windows because (predictably) it is the
platform on which we have to do the most gymnastics to get this to work.