python: filter versions less than 3.8 from interpreter dropdown (#6433)
### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #3740 filters out unsupported (<3.8) Python versions from the `Python:
Select Interpreter` dropdown


### QA Notes

1. create a Python interpreter <3.8 (if you are a pyenv user, you can do
`pyenv install 3.6`)
2. go to `Python: Select Interpreter`
3. should not appear in dropdown