Add new `r.packageTestExplorer` command (#6419)
Addresses #6265

This PR adds a new, additional command for running R package tests in
the Test Explorer UI, and maps the existing keybinding (from RStudio) to
the new command:

![Screenshot 2025-02-20 at 7 31
44 PM](https://github.com/user-attachments/assets/cbb419ba-80dd-4f8f-a3e6-d4011eb22495)

@:r-pkg-development

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- New command to run R package tests in the Test Explorer UI. The old
command is still available and can be mapped to a preferred keybinding,
if desired. (#6265)

#### Bug Fixes

- N/A


### QA Notes

- In an R package, the existing keybinding
<kbd>Cmd</kbd>+<kbd>Shift</kbd>+<kbd>T</kbd> should now pop up the Test
Explorer UI and run tests there.
- Both the old and new commands should be runnable from the command
palette.
- Both the old and new commands should be runnable multiple times in a
row.
- Let's also check the command's behavior is sensible if you open and
close the workspace.

---------

Signed-off-by: Julia Silge <julia.silge@gmail.com>
Co-authored-by: Jennifer (Jenny) Bryan <jenny.f.bryan@gmail.com>