Pre package extensions and install them on startup for users (#6403
Addresses #6302, #6301, #6305, #5315

Introduces the concept of "bootstrapped extensions". Bundles the
following extensions as VSIX files with Positron:
- Black
- Posit Publisher
- Shiny(Workbench only)
- Quarto
- ms-toolsai.jupyter-keymap
- ms-toolsai.vscode-jupyter-cell-tags
- ms-toolsai.vscode-jupyter-slideshow
- ms-toolsai.jupyter
- ms-pyright.pyright
- ms-python.debugpy

When a user starts up Positron after first installing it or a version
change (upgrade or downgrade), Positron will install the bootstrapped
extensions for the users. Users will then be able to manage the
extensions (upgrade, downgrade, uninstall, disable) themselves. Positron
will always prefer whichever version of the extension in newer (bundled
or user). When a user downgrades the extension, Positron will attempt to
upgrade to the latest packaged version on the next upgrade, but not next
Positron launch.

I changed the approach slightly from my initial spec. The initial spec
built on top of the existing
[defaultExtensionInitializer](https://github.com/posit-dev/positron/blob/main/src/vs/code/electron-utility/sharedProcess/contrib/defaultExtensionsInitializer.ts).
This approach is heavily inspired by that class but provides the server
and desktop logic in
src/vs/platform/extensionManagement/node/positronBootstrapExtensionsInitializer.ts.

### Release Notes

#### New Features

- Packages several extensions that were previously shipped a built-in
extensions as user managed extensions that are installed on first
launch.

### QA Notes
@:web