block interpreters from Positron dropdown (#6525)
Should block <3.8 interpreters from being registered when found. It
seems to fix everything for me locally, I think this is the only place
where old interpreters could sneak by.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #3740 


### QA Notes

1. Create a Python interpreter <3.8 (if you are a pyenv user, you can do
pyenv install 3.6)
2. Fo to `Python: Select Interpreter`, should not appear in dropdown
3. Go to Positron interpreter selector, should not appear here either