create the bootstrap .version file in the extensions directory rather than the application directory (#6554)
@midleman and @testlabauto reported that their tests relying on
extensions were failing locally. This is because the bootstrapped
extension feature was saving `.version` to the application folder rather
than the user's extension folder, causing the bootstrapped extensions to
only be installed for the first user that launched the application.
Because the tests use a temporary user data directory per run, the
extensions would not be installed on subsequent runs.

This would've prevented extensions from being installed for multiple
users on Workbench as well.

We now create, look for, and update `.version` in the user's extension
directory rather that the application directory.

In the implementation, `storageFilePath` was unnecessarily being
declared as a class member so I fixed that as well.