Omit Shiny from bootstrapped extensions (#6562)
The Shiny extension 1.1.2 is missing a macOS binary, which is causing
builds to fail when we try to bundle it.

```
[VSCode-darwin-universal/Positron.app/Contents/Resources/app/extensions/bootstrap/posit.shiny-1.1.2.vsix]
  End-of-central-directory signature not found.  Either this file is not
  a zipfile, or it constitutes one disk of a multi-part archive.  In the
  latter case the central directory and zipfile comment will be found on
  the last disk(s) of this archive.
unzip:  cannot find zipfile directory in one of VSCode-darwin-universal/Positron.app/Contents/Resources/app/extensions/bootstrap/posit.shiny-1.1.2.vsix or
        VSCode-darwin-universal/Positron.app/Contents/Resources/app/extensions/bootstrap/posit.shiny-1.1.2.vsix.zip, and cannot find VSCode-darwin-universal/Positron.app/Contents/Resources/app/extensions/bootstrap/posit.shiny-1.1.2.vsix.ZIP, period.
```

We shouldn't even be trying to bundle the Shiny extension for MacOS; it
is happening because the mechanism for selecting extensions for
inclusion in Workbench vs. Desktop needs to be reworked (this change
also removes the existing mechanism).

The fix here is just to remove the extension from the set of
bootstrapped extensions. I'll open a separate issue to add it back for
Workbench.