Improves the implementation of ActivityItemStream for code clarity (#6624)
When the Positron Console was implemented, a decision was made to create
two types of activity item streams, `ActivityItemOutputStream` for the
`stdout` stream and `ActivityItemErrorStream` for the `stderr` stream.
At the time, this seemed like the best implementation and it resulted in
three classes:

`ActivityItemStream` base class which contains all the logic for an
activity item stream.
`ActivityItemOutputStream` subclass.
`ActivityItemErrorStream` subclass.

Basically, this meant that we were using the type system to keep track
of the type of an activity item stream.

After reviewing https://github.com/posit-dev/positron/pull/4848 and
https://github.com/posit-dev/positron/issues/4519, and thinking about
the problem and the implementation more deeply, this PR makes the
following adjustments:

1. Adds a new `ActivityItemStreamType` enumeration and adds a `type`
property to the `ActivityItemStream` class.
2. Improves how new `ActivityItemStream` objects are created by adding a
`clone` method.
3. Removes `ActivityItemOutputStream` subclass.
4. Removes `ActivityItemErrorStream` subclass.
5. Simplifies impacted code.