Fix variables sending update event on empty cell execution (#6645)
Addresses https://github.com/posit-dev/positron/issues/4277.

The problem is that IPython runs `post_run_cell` hooks but not
`pre_run_cell` hooks when executing an empty cell. The fix is to exit
early in both cases if the cell was empty.

I also made a couple test changes:

1. Separated the change detection tests such that we don't need the
complicated logic to determine whether it's `assigned` or `unevaluated`.
2. Wrapped a bunch of test cases in `pytest.param` which makes it easier
to browse in the test explorer UI.
3. Reset the `_` variable to `''` instead of deleting it after each
test, which more closely matches real behavior. That seemed to affect
the ordering of variables in certain tests which I updated too.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fixed Python Variables flashing to signify an update when running an
empty code cell (#4277).