bump wait for connection (#6658)
This tests occaionally fails on windows, due to an empty connection pane.

We previously put in a very small wait, I'm bumped it by 1 second to see if this helps.

If this continues to be an issue, we could also wrap that click in a retry.

## QA Notes
All test should pass. The offending test is already marked "critical" so the PR run will pick it up.

@:win