python: show interpreter path when python files opened (#6647)
User feedback was that it can be difficult to see the full path of the
active interpreter. This is partially fixed with the new multiconsole
work, which has a much wider display to see the full path on interpreter
selection. We still want a zero-click way to see the interpreter path
though, so I've reused the capability that is already available in the
Python extension.

Since we already have the interpreter name displayed in numerous
locations, I feel okay removing it here.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #3033 

#### Bug Fixes

- N/A


### QA Notes

![Screenshot 2025-03-05 at 12 38
53 PM](https://github.com/user-attachments/assets/e341cb45-bc49-4d8e-9d93-26f49a8a36e3)

With a Python file open, the path to the active interpreter should show
in the bottom right of the status bar

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>