Improve error handling for downloads (#6673)
This change addresses an issue observed in builds in which problems
downloading our binary dependencies do not necessarily cause the build
to fail. Instead, the build "succeeds" but the product is broken because
the dependencies are missing.

The fix is to add more error handling to the download scripts:

- make sure we got a 200 status code on our final request for the binary
data (after following 302 redirects)
- make sure we got a reasonable-looking amount of data for the asset

Addresses https://github.com/posit-dev/positron/issues/6672