Update `http-proxy-middleware` (#6728)
Getting the latest `http-proxy-middleware` updates it and several other
packages. The major version change does require updating how the proxy
is constructed since the signature changed.

The legacy constructor is used but with the new options format. Using
the new constructor doesn't seem to work but we need these updates.