upstream merge for vscode python `v2025.2.0` (#6722)
[Release notes for
v2025.2.0](https://github.com/microsoft/vscode-python/releases/tag/v2025.2.0).

Not too much here that will affect us. The primary work i:
- integration with the Python environments extension 
- updates to Native REPL
- telemetry cleanup 
 
Nearly all of the changed lines are from the `package-lock.json`.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

Maintenance #6581 

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

---------

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>
Co-authored-by: seem <mwlorgat@gmail.com>