Data Explorer: Support pandas.CategoricalDtype, polars.Categorical, enable sparklines for Polars (#6739)
Addresses #3417 for Python. Categorical/Factor data types should return
their schemas as the display type of the categories, which for R and
Polars should be string, but for pandas it is whatever the display type
of the dtype's categories member is.

I also noted that sparklines were disabled for Polars, but they had unit
tests so I enabled them and they should work now!