Restore console history at startup (#6741)
This change persists the contents of the Console tab, so that reloading
the window or resuming your session on Posit Workbench restores the
history of the commands you have executed and their results.


https://github.com/user-attachments/assets/2e1118ce-7cec-47b0-9175-de0a5733b5dd

A few related improvements are made, too:

- Now that we have console contents, we show them right away at startup
when reloading or reconnecting -- even before extensions are started,
etc. (see video) This replaces the flickering "starting..." or
"preparing..." screens and makes reconnection feel snappier.
- Input history has been revamped; we now have separate histories for
navigation inside a session (e.g. up/down) vs. searching (e.g. cmd/ctrl
up). The former will take you to literally the previous thing you typed
in that session's window, the latter to all the code you've executed in
the matching language.

Under the covers, the execution history is stored and recorded
separately from the console, by a separate set of listeners to the
messages emitted from the runtimes.

Addresses #4925. 

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Console contents are retained when reconnecting to existing sessions
(#4925)

#### Bug Fixes

- N/A


### QA Notes

The contents are stored in a format that is somewhat optimized for speed
and compactness, and is intended to preserve the record of code you have
run and its output, rather than faithfully re-create the exact console
contents before reloading. For example:

- Console events that are not executions or outputs (e.g. restart
notifications, etc.) are not retained.
- Standard error and standard output are treated identically, so red
text could become plain after reloading.
- Only plain text output is retained.
- Only one error per code execution event is retained.

Tests: `@:console`