Update foreground session when interpreter changes (#6756)
Addresses #6719 

This fixes a regression introduced to `$selectLanguageRuntime` when
multiple console sessions is enabled.

The `$selectLanguageRuntime` function exposed to the extension host by
Positron is meant to be used by a language pack to signal to Positron
that the interpreter/runtime has changed on their end and that Positron
should take appropriate action. Right now that means Positron should
change the foreground session to a session for that runtime.

The foreground session is not updated by Positron when
`$selectLanguageRuntime` is being called which is causing the user to
see mismatched information in the UI.

The fix is to update the foreground session in Positron when
`$selectLanguageRuntime` is called.

Note: This doesn't address the inverse path. When the foreground session
changes, the Python extension does not currently change the active
interpreter. That can be handled after
https://github.com/posit-dev/positron/pull/6714 which provides a way for
extensions to be notified the foreground session has changed.

### Release Notes


#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:sessions @:console @:web

### Screenshots

https://github.com/user-attachments/assets/59b878a1-87e1-48d7-9307-b8a1ce7421e2