e2e-test: gracefully handle repo cloning when offline (#6771)
### Summary
- Updated the script in our e2e tests that clones `qa-example-content`
to gracefully handle the scenario when user is offline (for example, on
a plane headed to/from work week). Previously user was unable to launch
playwright due to cloning error.

### QA Notes

Feel free to test it out locally with by toggling WIFI on and off and
running either of these commands:
* `FORCE_CLONE=true npx playwright test fast-execution --project
e2e-electron`
* `FORCE_CLONE=false npx playwright test fast-execution --project
e2e-electron`