multiple console session: start and stop LSPs (#6714)
Addresses #6310 

To get LSP features working in a sensible manner for multiple console
sessions we need to ensure the LSP for the active session is the only
one providing diagnostics/code completion/etc.

The simplest approach to achieve the above is to only allow the LSP for
the active session to be running, and shutdown all other LSPs. Positron
does not know about the LSPs, these are managed by the extensions right
now. To keep with this pattern, we need to let the language pack know
when the active session changes so it can gracefully start/stop the LSPs
for each session.

When a user changes the active session, we fire an
`onDidChangeForegroundSession` event in Positron. By exposing this event
to the extension host, the language packs can listen for changes to the
active session and start/stop the LSPs.

We can notify the extension host when this event fires and the extension
host can mirror the event and emit when its notified the active session
changed. The mirrored event on the extension host side can be exposed in
the positron API so that an extension can listen to the event. The
extensions can listen to this event and start the LSP for the active
session, and stop all other LSPs.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:console @:sessions @:web @:editor

---------

Co-authored-by: Sam Clark <sam.clark@posit.co>
Co-authored-by: Julia Silge <julia.silge@gmail.com>