Allow users to specify a bootstrap extensions directory on server startup (#6814)
Addresses #6304 

Allow users to specify an `--bootstrap-extensions-dir` on server startup
and install all VSIX files from this folder on launch. Like bootstrap
extensions, this will only happen after a version upgrade so that users
ultimately control their own extensions.

This feature is primarily for Workbench so release notes will be
documented as part of Workbench.

### QA Notes

@:web-only