Make positron-python CI use uv (#6797)
Addresses another idea in #6290.

This PR changes the `positron-python-ci` and `positron-python-nightly`
workflows to use `uv` for managing python and dependencies. This was a
pretty quick change so I'm fine if we don't think it's a good idea. I
think most jobs are about twice as fast now (and some Windows jobs are
even faster).

What this PR doesn't do yet:
- touch the `typescript-tests` job, because of the custom venv setups in
there
- manage the `pinned-test-requirements.txt` file, though I think this
could be better with `uv` in the future
- manage installing dependencies with `uv` in other scripts, like when
we vendor or bundle