Implement JEP 66 to fix/mitigate socket allocation race condition (#6874)
This change brings in a new version of the kernel supervisor that
implements [JEP
66](https://jupyter.org/enhancement-proposals/66-jupyter-handshaking/jupyter-handshaking.html).
This should prevent the socket allocation race condition when using the
Ark kernel, which [implements the other
side](https://github.com/posit-dev/ark/pull/576) of JEP 66.

Unfortunately ipykernel does not appear to support this proposal yet.
However, the supervisor also has an improvement for kernels that don't
support JEP 66: before this change, sockets were allocated when a
session was created, and then used when the session was started. Because
these are two different API calls, there's some opportunity for the
sockets to get bound by other software between the two calls. After the
change, creating a session does not allocate sockets; they are allocated
immediately before the session is started in the same API call.

Addresses https://github.com/posit-dev/positron/issues/5799.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Improve R and Python session startup reliability by changing the way
ports are negotiated (https://github.com/posit-dev/positron/issues/5799)


### QA Notes

- Restarting is also affected by this code.
- You can check the Positron Supervisor logs to see if this change is
working. You should see JEP 66 referenced when the an R session starts,
but not when a Python session starts.
- After this change, R should not experience any more "address in use"
errors.
- This change should reduce (and maybe even eliminate) Python "address
in use" errors, but it isn't guaranteed to do so since it only reduces
rather than eliminates the window of time between socket allocation and
socket binding.

Test tags `@:console`