python: changed order in test-requirements.txt (#6899)
Addresses #6890.

When we switched to installing dependencies with `uv`, the resolver
logic slightly changed from the old pip logic, which ended up giving us
different versions of various packages than before (see
https://github.com/posit-dev/positron/issues/6890#issuecomment-2741080283).
In particular, we landed in a state where the versions of `bokeh`,
`holoviews`, and `panel` were incompatible in practice, despite their
metadata saying that they _were_ compatible.

After some experimentation, I found that if we move `bokeh` to the end
of the file, then the `uv` resolver will try to solve it last. This is
good because by that point, the picky `holoviews` is already solved, and
we can just install a compatible `bokeh` version.

Sorry I forgot to run the nightly tests before merging that PR! Will do
before I merge this one 🙂