Let the Jupyter Kernel pick the DAP/LSP server port (#6846)
Requires https://github.com/posit-dev/ark/pull/750

This PR is in the same realm as #6874 in that it remove the potential
for a race condition in our TCP connection between the LSP client and
server. #6874 is about removing the potential for a race condition in
the core jupyter kernel sockets (iopub, heartbeat, etc). Remove this LSP
related race condition is going to be important in multi-session world,
because each time we switch to a new session we pick a new port, which
otherwise would greatly increase the risk of port related race
conditions.

There are now two messages used in the LSP/DAP Comm startup sequence.
I'll use Rust to describe them:

```rust
/// Message sent from the frontend requesting a server to start
pub struct ServerStartMessage {
    /// The IP address on which the client is listening for server requests. The server is
    /// in charge of picking the exact port to communicate over, since the server is the
    /// one that binds to the port (to prevent race conditions).
    ip_address: String,
}

/// Message sent to the frontend to acknowledge that the corresponding server has started
pub struct ServerStartedMessage {
    /// The port that the frontend should connect to on the `ip_address` it sent over
    port: u16,
}
```

- Positron sends the kernel an `ip_address` (right now it is always
`127.0.0.1`) and then waits to receive a `port`
- The kernel does whatever it needs to bind to a port on that ip
address. For ark this looks like trying to connect to `127.0.0.1:0` and
letting the OS assign us a port.
- The kernel sends back the `port` and the server side of the LSP/DAP
starts listening for connections on that port.
- Positron gets the `port` and finishes out the handshake by setting up
the LSP client and performing a `socket.connect(port)`, which also
unblocks the server since it has been listening for that connection



https://github.com/user-attachments/assets/6bec3556-f95c-4da8-af0f-37a4c68013f7

`@:web` `@:console` `@:sessions` `@:notebooks`

---------

Signed-off-by: Sam Clark <samclark2015@users.noreply.github.com>
Co-authored-by: Sam Clark <samclark2015@users.noreply.github.com>