Updates Console to better deal with large amounts of output. (#6912)
This PR addresses #5441 by implementing a "truncate at top" scrollback
strategy in the Console component. Additionally, it adds two new
features to the console: 1) the ability to control how much output will
appear in the console (see #6913), and 2) the ability to open the entire
Console history in an Editor (see #6914).