Only render Console instances when the `consolePaneWidth` is greater than 0 (#6973)
This addresses #6845 by only rendering `ConsoleInstance` components
when the `consolePaneWidth` is greater than 0. (There's no need to
render a zero-width component since it will render nothing.) This
prevents the `ConsoleInput` component from setting the
`CodeEditorWidget` component's width to something that's too small,
which causes it to behave badly. (Ultimately, this is a
`CodeEditorWidget` bug that does not have an easy or straightforward
fix.)