Reactivate the LSP of restarting sessions (#6938)
Addresses https://github.com/posit-dev/positron/issues/6902, joint work
with @seeM
Attempt 2 replacing https://github.com/posit-dev/positron/pull/6924 
Joint PR with #6947 

As noted in https://github.com/posit-dev/positron/pull/6924, restarting
sessions won't have `onDidChangeForegroundSession` events fired for
them, because the foreground session isn't changing. And we think this
is correct because many things rely on the foreground session event not
firing unnecessarily.

Instead, it seems like it is up to the language extension to track when
the LSP needs to come online. This isn't super easy, but I've come up
with:

- When a session becomes the R foreground session through
`onDidChangeForegroundSession`, activate the LSP and deactivate all
others.

- When a session restarts AND reenters the `Ready` state AND was
previously the R foreground session, activate the LSP and deactivate all
others. We are not going to get a `onDidChangeForegroundSession`
notification here, so it is up to us to handle this case.

Implementing the 2nd bullet was a bit gross. I've chosen to do it in
`RSessionManager` for now, but I think that is going to eventually get
absorbed into `RRuntimeManager` (that's not that important to the PR,
just a note).

---

An important note about this PR is that it fully transitions ownership
over _starting_ the LSP from `RSession` itself up into
`RSessionManager`. In a multi console world, it is impossible for
`RSession` itself to correctly start its own LSP because it has to wait
for all other LSPs to shut down first - something only a session manager
can do.

@:web @:sessions @:notebooks