Teach Positron Assistant to work with R packages (#6954)
This is a proof of concept that shows how language packs can provide
tools to the model. It adds tools from the R language pack that give the
model the ability to:

- Check the version of R packages that are installed
- See which R packages are loaded
- Install new R packages, with user confirmation

Notes:

- Requires https://github.com/posit-dev/ark/pull/752 on the backend
- The package installation flow doesn't feel amazing; I think it needs
to work more interactively (i.e. send `install.packages(....)` to the
console and read the output). That isn't possible with today's API but
I'll work on making it so.