Add Support for WSL in Positron Launcher (#6941)
#### New Features

- This PR updates the WSL launcher script (`code.sh`) to support
[kv9898.open-remote-wsl](https://github.com/kv9898/open-remote-wsl), a
community-maintained Positron-compatible WSL extension. This allows
users to open WSL directories in Positron from WSL command line:

```sh
positron .
```

This mimics the WSL "code ." feature from VSCode and the official
microsoft WSL extension. Note that even VSCodium does not have this
feature.

#### Bug Fixes

- N/A


### QA Notes

The `resources\win32\bin\code.sh` file is for dealing with WSL - and the
Positron inherits this script from CodeOSS. Nevertheless, the original
scripted is adapted for the official "ms-vscode-remote.remote-wsl"
extension, which is not usable by Positron. This PR adapts the script by
trying to find the alternative "kv9898.open-remote-wsl" extension, and
launch the WSL extension.

Note that the WSL extension automatically downloads the Positron server
at the first connection. This requires additional information such as
Positron's version and build number to retrieve the correct download
URL.