Add code execution support to Positron Assistant (#7006)
This change adds code execution capabilities to Positron Assistant. This
allows Assistant to run R or Python code repeatedly to accomplish a task
give by the user. The user must review and confirm each code execution.

Code is executed in the Console and becomes part of the execution
history as though the user had entered it themselves. The results of the
code are both displayed to the user and sent back to the model.


![image](https://github.com/user-attachments/assets/d31c2d55-621f-413c-9cc8-30d8cc6fb105)

There are a few rough edges here; the biggest one is that there are now
two competing ways for the model to return code, each with their own
tradeoffs -- i.e. does the user want the model to _generate_ code (to
ultimately insert into a script, run/tweak, etc.) or _run_ code and
iterate on it in a more agentic way? In my experimentation the model was
pretty good at deciding which workflow to use, so I think this is a good
start, but we also will want to think about how we can better unify
these experiences.

Most of the code here is in support of an overhaul of the `executeCode`
API. This API now accepts an `ExecutionObserver` that gets notified in
real time about output and relevant execution status events, and has a
cancellation token to request interruption. The API also now returns a
promise that resolves with the execution result or rejects if execution
fails.

Addresses #6968.
Addresses #2622.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>