Assistant: Update to new code editing API (#7032)
The code editing provider API has changed to support mapped edits that
span several documents, and to stream the results instead of waiting for
them at the end.

This PR makes the minimal change necessary to restore the functionality
Assistant had prior to the change by implementing the new API.

Addresses #6960.