Assistant: Execute code in the correct console (#7030)
This small change causes code run from the "Run in Console" gesture in
Assistant to run in a console of the correct (matching) language instead
of the active console.

Addresses #6962. 

## QA Notes

Another side effect of this change is that if you run code but don't
have a console running for the matching language, one should be started
for you.