fix isCustomEnvironment check
Check if any custom environment directories are parents of the interpreter path, instead of just checking if the interpreter path is specified in the include or override settings.

This is needed because the resolved interpreter path may not be the same as the one specified in the settings.

e.g. ~/scratch/3.10.4/bin/python3 may be resolved to ~/scratch/3.10.4/bin/python because we prioritize the basic python executable.